Implemented the server and workers
This implements the basic server. It listens for connections and hands them off
to the workers. The workers spins up a new thread that handles the connection.
The server's loop of listening for connections is run in a separate thread,
allowing the server to still be interacted with after it's started.