Added class documentation and cleaned up a few features
This adds class documentation and explanation of implementation details. This
also cleans up and expands on a few features. I added a logger abstraction to
handle wrapping different loggers and to handle when no logging is wanted.
Additionally, I added a socket wrapper to provide convenience methods for
working with the socket.