Added tests, modified documentation and implementation slightly because of them
Added unit and system tests to cover the features of the server. Additionally
added a `ready_timeout` option for controlling the responsiveness of the server.
Also implemented a mutex and condition variable for locking accessing and
modifying the server's thread. This allows the `stop` method to wait for the
server thread to finish.