Allow building a DatTCP server from file descriptors
This allows both creating the TCP server from a file descriptor
and passing in client socket file descriptors and having those
loaded onto the worker pool. With this a server process can hand
off it's TCP server and client socket file descriptors to another
process, essentially to pick up where it left off. This is to
support a "hot restart" when using a DatTCP server.