Change `Server` to a class instead of a mixin
This makes `Server` a class instead of a mixin. This makes
DatTCP more congruent with DatWorkerPool and should be a simpler
design. This avoids method collision and a design of overriding
methods.

This disables some of the test suite for future changes. The
callback tests are commented out as we are planning on changing
them. With the change to a class, the callbacks aren't really
usable.

Closes #22