Don't validate that min workers is less than max workers
This removes the validation for a DatTCP server that would ensure
the min workers is less than the max workers. With DatWorkerPool
being it's own gem, this is it's responsiblity and not DatTCPs.
Futhermore, DatTCP/DatWorkerPool will run if the min workers are
more than the max workers. The behavior may be strange, but if
you configure it with bad data you should get bad behavior.

Apparently we were not testing this feature either, which is
another good reason to remove it.

Closes #24