Change the `wait_for_shutdown` default from true to false
The main reasoning for originallying defaulting it to true was
because of the test suite (where we typically want to ensure a
server shuts down before moving on). This is not a good use-case
for a default though. The more common case in actual usage is to
not wait for shutdown. This was evident when writing `Sanford`
and while working on `Qs`. This lines up this behavior so it is
consistent with how `Qs` works.