Remove `accept_connection` loop, not needed
This removes the while loop in the `accept_connection` method.
This is ultimately not-needed and just creates extra overhead.
I'm not sure why I originally thought this was needed, but it's
a loop within a loop that both have the same conditions. This
might have been due to legacy concerns. This is a simpler
implementation that achieves the same functionality.