Remove callbacks/hooks, keep configuring `TCPServer` ability
This removes the callbacks/hooks from DatTCP. There is almost no
need for these currently and their previous implementation doesn't
work with the `Server` being a class instead of a module.

The `configure_tcp_server` callback is gone, but it's been
replaced by calling `listen` with a block. The `listen` method
will yield the `TCPServer` instance to the block. This allows
setting socket options before the `TCPServer` begins listening
for new connections.

Closes #21