Reset benchmarks to new baseline
This resets the benchmarks to a new baseline. I've switched from
the computer I used to use to develop DatTCP so the benchmarks
don't line up with the old one. This runs them on the new machine
with no other changes to set a new baseline to compare changes
to.

/cc @kellyredding