Don't use Rake for the benchmarking scripts
This switches from using Rake to run the benchmark scripts to
just using ruby script files. This minimizes using Rake with the
gem and only requires it for gem building. I'd rather not rely on
extra dependencies that are not needed and for this case, Rake
isn't needed.

This removes the "run X requests against the Benchmark server"
functionality. This hasn't been used since DatTCP has been stable
and isn't related to "benchmarking". It was more of a system test
to see that everything worked for an echo server.