Change the ip method to return the socket ip not its hostname
This changes the ip method to return the socket ip not its
hostname. This more closely matches the method name. A hostname
can still be used when calling `listen`.

This showed up in Sanfords test suite. I wanted to switch to using
`127.0.0.1` instead of `localhost` because there are situations
where a computer has trouble looking up an ip for a hostname and
that causes the test suite to fail. I couldn't switch everything
though because one of the tests expected the ip being passed to
`listen` to be returned with the `ip` method but it returned
`localhost`. This confused me because the method is called `ip`
so I didn't immediately understand why it would return the hostname
instead.