Update to use the new dat-worker-pool
This updates dat-tcp to use the new dat-worker-pool.
dat-worker-pool's interface has changed but functionally it works
the same as it previously did.

When a server is initialized it now takes a worker class and a
set of options instead of just taking options. This is because
dat-worker-pool now requires a worker class when its initialized.
A block is no longer part of the server or dat-worker-pool's
interface. The logic for processing a work item (a client socket
for dat-tcp) is now part of the worker class that is passed in.
The server also no longer takes worker callbacks, these should be
added to the worker class thats passed in.

The server now also builds a worker pool when its initialized
instead of waiting till its started. This was previously made
possible when dat-worker-pool was changed to not start
automatically but dat-tcp didn't take advantage of it til now.
This avoids some conditionals and having to set/unset the worker
pool ivar as the server is started/stopped.

As part of this change, dat-tcp now also provides a `Worker` mixin
which for now only brings in dat-worker-pools `Worker` mixin. It
also requires the worker class that is passed to the server use
this mixin.

This also includes a number of small changes related to using the
new dat-worker-pool and the server interface changing:

* The server no longer automatically closes the client socket. This
  is alittle aggressive for something as generic as dat-tcp. Not
  all servers will always want to do this and it doesn't work
  nicely with the new worker class pattern of dat-worker-pool.
* The server no longer takes a min and max number of workers
  because dat-worker-pool no longer supports a min or max. It now
  defaults to 2 workers (its previous min value).
* Renamed the passed in client file descriptors. This confused me
  because the local variable and the method were the same name.
* Removed the thread status from the inspect. I'd rather not expose
  this level of detail about how the server works.
* Changed the work loop organization to try and make it more
  readable. I didn't find it very approachable when I was trying to
  understand its logic.
* Updated styles and conventions to our latest standards
  throughout.