Hotfix: Reset benchmarks to current machine
This re-runs the benchmarks to reset them to match my current
machine. This is so future benchmarks can be compared without the
change in machine skewing the results.

Side-note: Not sure why the benchmarks are running slower. My
current machine should be the same as it was the last time I ran
the benchmarks except it has a SSD drive (which should in-general
make things faster) and I've updated the OS to Yosemite (I believe
I was previously on Mountain Lion).

/cc @kellyredding