Change bench script to only be a single command
This simplifies the bench script so it only requires running a
single command instead of 2 separate commands. This makes it more
convenient to run. This also makes the bench script as consistent
as possible with qs.

This brings in scmd and uses it to spin up the "server" script from
the "report" script. The "server" script can still be spun up
independently for testing if needed.