Allow passing a logger to a server, remove debug option
This updates the server to take a logger option. This replaces the
debug option and will turn on logging for dat-tcp and its worker
pool. This provides more flexibility for logging and doesn't force
only logging to stdout.

The server now only logs if an error occurs or when it closes its
tcp server. This is because it now passes its logger to
dat-worker-pool which handles logging that the worker pool is
starting or shutting down. Both dat-tcp and dat-worker-pool logging
is debug logging and in general shouldn't be enabled in a live
environment.