Update server spy to new server interface
This updates the server spy to match the new server interface. This
is part of making sure the server spy is a good spy and useful. The
server spy now takes a worker class and makes sure it is a dat-tcp
worker. It also no longer takes any worker procs.

This also adds default constants for the server configuration. This
is so the server spy and use the same defaults.