Rename `Signal` to `State`, use dat-worker-pool locked object
This renames the `Signal` class to `State` updates it to inherit
from dat-worker-pool's locked object. This also updates the state
handling to be more consistent with qs.

This renames the `Signal` class to `State` which better matches
its current usage. When `pause`, `stop` or `halt` are called they
now immediately set the state and write a "signal" to the signal
writer IO. The signal is just a simple value to wakeup the work
loop thread. Previously, the signal would also set the state. This
simplifies the logic and makes its behavior consistent with qs.

The `State` class now also inherits from dat-worker-pool's locked
object. This handles the mutex locking and now the `State` class
only implements an interface for checking what the current state
is.

This also updates the exception handling for the work loop to set
the state to "stop". This ensures it shuts down correctly and that
the state variable matches the current state of the server.